Finished code generation functions; Created Code class
Finished the code generation functions, and wrapped it in a Code class for simplicity. All functions needed are public functions.

The code still needs to be refactored, though. It could be simpler and two functions are remarkably similar, and I will merge them at some point, but it won't affect the way to interact with the class.